requirements.txt
feedparser==6.0.8
requests>=2.28.2,<3.0.0
Pillow>=9.4.0,<10.0.0
tkhtmlview==0.2.0
